requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
duckduckgo-search==3.8.5
exceptiongroup==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.0.292
langsmith==0.0.92
lxml==5.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.9.0
numpy==1.26.4
openai==0.28.0
packaging==23.2
pandas==2.2.1
Pillow==9.5.0
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.0
Pygments==2.17.2
Pympler==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.29
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.1
validators==0.24.0
watchdog==4.0.0
wikipedia==1.4.0
yarl==1.9.4
zipp==3.18.1
